hello-world-stack/backend/requirements.txt
Django>=5.1.4
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
